+++ src/Symfony/Component/DomCrawler/Test/Constraint/CrawlerSelectorCount.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\DomCrawler\Test\Constraint;
 
 use PHPUnit\Framework\Constraint\Constraint;
-use Symfony\Component\DomCrawler\Crawler;
+use Symfony\Component\DomCrawler\DomCrawler;
 
 final class CrawlerSelectorCount extends Constraint
 {
@@ -28,7 +28,7 @@ final class CrawlerSelectorCount extends Constraint
     }
 
     /**
-     * @param Crawler $crawler
+     * @param DomCrawler $crawler
      */
     protected function matches($crawler): bool
     {
@@ -36,7 +36,7 @@ final class CrawlerSelectorCount extends Constraint
     }
 
     /**
-     * @param Crawler $crawler
+     * @param DomCrawler $crawler
      */
     protected function failureDescription($crawler): string
     {