+++ src/Symfony/Component/DomCrawler/Test/Constraint/CrawlerSelectorTextSame.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\DomCrawler\Test\Constraint;
 
 use PHPUnit\Framework\Constraint\Constraint;
-use Symfony\Component\DomCrawler\Crawler;
+use Symfony\Component\DomCrawler\DomCrawler;
 
 final class CrawlerSelectorTextSame extends Constraint
 {
@@ -31,7 +31,7 @@ final class CrawlerSelectorTextSame extends Constraint
     }
 
     /**
-     * @param Crawler $crawler
+     * @param DomCrawler $crawler
      */
     protected function matches($crawler): bool
     {
@@ -44,7 +44,7 @@ final class CrawlerSelectorTextSame extends Constraint
     }
 
     /**
-     * @param Crawler $crawler
+     * @param DomCrawler $crawler
      */
     protected function failureDescription($crawler): string
     {