+++ src/Symfony/Component/DomCrawler/Tests/Test/Constraint/CrawlerSelectorExistsTest.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\DomCrawler\Tests\Test\Constraint;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
 use PHPUnit\Framework\TestFailure;
-use Symfony\Component\DomCrawler\Crawler;
+use Symfony\Component\DomCrawler\DomCrawler;
 use Symfony\Component\DomCrawler\Test\Constraint\CrawlerSelectorExists;
 
 class CrawlerSelectorExistsTest extends TestCase
@@ -22,12 +22,12 @@ class CrawlerSelectorExistsTest extends TestCase
     public function testConstraint()
     {
         $constraint = new CrawlerSelectorExists('title');
-        $this->assertTrue($constraint->evaluate(new Crawler('<html><head><title>'), '', true));
+        $this->assertTrue($constraint->evaluate(new DomCrawler('<html><head><title>'), '', true));
         $constraint = new CrawlerSelectorExists('h1');
-        $this->assertFalse($constraint->evaluate(new Crawler('<html><head><title>'), '', true));
+        $this->assertFalse($constraint->evaluate(new DomCrawler('<html><head><title>'), '', true));
 
         try {
-            $constraint->evaluate(new Crawler('<html><head><title>'));
+            $constraint->evaluate(new DomCrawler('<html><head><title>'));
         } catch (ExpectationFailedException $e) {
             $this->assertEquals("Failed asserting that the Crawler matches selector \"h1\".\n", TestFailure::exceptionToString($e));